datagen.c
This is a multi-threaded c program to generate 0-byte files in nested directories.  The files generated have different file extensions which can be useful for testing I/O and data tagging. Change the parameters to scale out the datagen.  By default this program creates a directory called "manyfiles" and underneath that is creates thread directories based on the thread count defines and sub directories dir0-dirx with each directory having x number of files of different extensions.